separate `core` and `mods` from derived `tokens`
